V0.9.5.0：近期工单问题反馈修复

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/windows/WndGoShop.java
@@ -80,8 +80,10 @@ public class WndGoShop extends Window {
 
                     ArrayList<Ankh> ankh = hero.belongings.getAllItems(Ankh.class);
                     for (Ankh w : ankh.toArray(new Ankh[0])){
-                        Dungeon.level.drop(w, hero.pos).sprite.drop();
-                        w.detachAll(hero.belongings.backpack);
+                        if(!w.spankh){
+                            Dungeon.level.drop(w, hero.pos).sprite.drop();
+                            w.detachAll(hero.belongings.backpack);
+                        }
                     }
 
                     Statistics.fireGirlnoshopping = true;
